simplified custom_sell

+++ vin.py
@@ -21,8 +21,7 @@ class ViN(IStrategy):
     f_trades = './user_data/vintrades.txt'
     min_candle_vol: int = 0
     custom_buy_info = {}
-    stoploss_count: int = 0
-    sideways_candles: int = 36
+    max_concurrent_buy_signals_check = False
 
     minimal_roi = {"0": 100}
     stoploss = -0.99
@@ -44,13 +43,11 @@ class ViN(IStrategy):
     def populate_indicators(self, df: DataFrame, metadata: dict) -> DataFrame:
         df['green'] = (df['close'] - df['open']).ge(0)
         df['bodysize'] = (df['close'] / df['open']).where(df['green'], df['open'] / df['close'])
-        # df['uppertail'] = (df['high'] / df['close']).where(df['green'], df['high'] / df['open'])
-        # df['lowertail'] = (df['open'] / df['low']).where(df['green'], df['close'] / df['low'])
         hi_adj = df['close'].where(df['green'], df['open']) + (df['high'] - df['close']).where(df['green'], (df['high'] - df['open'])) / df['bodysize']
         lo_adj = df['open'].where(df['green'], df['close']) - (df['open'] - df['low']).where(df['green'], (df['close'] - df['low'])) / df['bodysize']
         df['hlc3_adj'] = (hi_adj + lo_adj + df['close']) / 3
         df['lc2_adj'] = (lo_adj + df['close']) / 2
-        # df['hc2_adj'] = (hi_adj + df['close']) / 2
+        df['hc2_adj'] = (hi_adj + df['close']) / 2
         df[f"candle_count_{self.startup_candle_count}"] = df['volume'].rolling(window=self.startup_candle_count, min_periods=self.startup_candle_count).count()
         df_closechange = df['close'] - df['close'].shift(1)
         s = (1, 2, 3)
@@ -60,16 +57,14 @@ class ViN(IStrategy):
         df['streak_s_min'] = df[[f"streak_{i}" for i in s]].min(axis=1)
         df['streak_s_min_change'] = df['close'] / df['close'].to_numpy()[df.index.to_numpy() - df['streak_s_min'].abs().to_numpy()]
         df['streak_s_max'] = df[[f"streak_{i}" for i in s]].max(axis=1)
-        # df['streak_s_max_change'] = df['close'] / df['close'].to_numpy()[df.index.to_numpy() - df['streak_s_max'].abs().to_numpy()]
         df.drop(columns=[f"streak_{i}" for i in s], inplace=True)
         df = self.populate_indicators_buy(df, metadata)
         df = self.populate_indicators_sell(df, metadata)
-        if self.config['runmode'].value not in ('live', 'dry_run'):
-            if self.write_to_csv:
-                ef = df[['date', 'open', 'high', 'low', 'close', 'volume']].reset_index()
-                ef['pair'] = metadata['pair']
-                with open(self.df_csv, 'a') as f:
-                    ef.to_csv(f, sep=';', header=f.tell()==0, index=False)
+        if self.config['runmode'].value not in ('live', 'dry_run') and self.write_to_csv:
+            ef = df[['date', 'open', 'high', 'low', 'close', 'volume']].reset_index()
+            ef['pair'] = metadata['pair']
+            with open(self.df_csv, 'a') as f:
+                ef.to_csv(f, sep=';', header=f.tell()==0, index=False)
         return df
 
     def fill_custom_buy_info(self, df:DataFrame, metadata: dict):
@@ -107,32 +102,31 @@ class ViN(IStrategy):
     def confirm_trade_entry(self, pair: str, order_type: str, amount: float, rate: float,
                             time_in_force: str, current_time: datetime, **kwargs) -> bool:
         df, _ = self.dp.get_analyzed_dataframe(pair, self.timeframe)
-        candle_1: Series = df.iloc[-1]
+        candle_1 = df.iloc[-1]
         buy_candle_date = candle_1['date']
-
-        if buy_candle_date in self.custom_buy_info.keys():
-            pairs = len(self.dp.current_whitelist())
-            max_concurrent_buy_signals = max(int(pairs * 0.08), 10)
-            buy_tags = self.custom_buy_info[buy_candle_date][pair]
+        try:
             buy_info = self.custom_buy_info[buy_candle_date]
+            buy_tags = buy_info[pair]
             buy_signal_count = buy_info['buy_signals']
-            if buy_signal_count > max_concurrent_buy_signals:
-                log.info(f"confirm_trade_entry: Buy for pair {pair} with buy tag {buy_tags} on candle {buy_candle_date} is cancelled. There are {buy_signal_count} concurrent buy signals (max = {max_concurrent_buy_signals}).")
-                return False
-            else:
-                log.info(f"confirm_trade_entry: Buy for pair {pair} with buy tag {buy_tags} on candle {buy_candle_date}.")
-                if self.config['runmode'].value not in ('live', 'dry_run') and self.write_to_txt:
-                    close_1_price = candle_1['close']
-                    indicator = []
-                    i = 16
-                    indicator.append(candle_1[f"mom_{i}"])
-                    indicator.append(candle_1[f"mom_{i}_low"])
-                    indicator.append(candle_1[f"mfi_{i}"])
-                    indicator.append(candle_1[f"close_corr_{i}"])
-                    with open(self.f_buys, 'a') as f:
-                        print(f"{pair};{buy_candle_date};{rate:.10n};{buy_tags};{close_1_price:.10n}", *indicator, sep=';', file=f)
-                return True
-        else:
+            if self.max_concurrent_buy_signals_check:
+                pairs = len(self.dp.current_whitelist())
+                max_concurrent_buy_signals = max(int(pairs * 0.08), 10)
+                if buy_signal_count > max_concurrent_buy_signals:
+                    log.info(f"confirm_trade_entry: Buy for pair {pair} with buy tag {buy_tags} on candle {buy_candle_date} is cancelled. There are {buy_signal_count} concurrent buy signals (max = {max_concurrent_buy_signals}).")
+                    return False
+            log.info(f"confirm_trade_entry: Buy for pair {pair} with buy tag {buy_tags} on candle {buy_candle_date}.")
+            if self.config['runmode'].value not in ('live', 'dry_run') and self.write_to_txt:
+                close_1_price = candle_1['close']
+                indicator = []
+                i = 12
+                indicator.append(candle_1[f"mom_{i}"])
+                indicator.append(candle_1[f"mom_{i}_low"])
+                indicator.append(candle_1[f"mfi_{i}"])
+                indicator.append(candle_1[f"close_corr_{i}"])
+                with open(self.f_buys, 'a') as f:
+                    print(f"{pair};{buy_candle_date};{rate:.10n};{buy_tags};{close_1_price:.10n}", *indicator, sep=';', file=f)
+            return True
+        except:
             log.warning(f"confirm_trade_entry: No buy info for pair {pair} on candle {buy_candle_date}.")
             return False
 
@@ -167,31 +161,25 @@ def vwrs(df: DataFrame, length: int) -> Series:
 
 
 class ViNBuyPct(ViN):
-    buy_lookback_range = range(12, 33)
+    buy_lookback_range = range(10, 33)
     def populate_indicators_buy(self, df: DataFrame, metadata: dict) -> DataFrame:
         df_closechange = df['close'] - df['close'].shift(1)
-        i = 16
+        i = 12
         df['updown'] = np.where(df_closechange.rolling(window=i, min_periods=2).sum().gt(0), 1, np.where(df_closechange.rolling(window=i, min_periods=2).sum().lt(0), -1, 0))
         df[f"streak_b"] = df['updown'].groupby((df['updown'].ne(df['updown'].shift(1))).cumsum()).cumsum()
         df.drop(columns=['updown'], inplace=True)
-        # ef = df['close'].reset_index()
         for i in self.buy_lookback_range:
-            # j = i * 2
             df[f"pctchange_{i}"] = df['close'].pct_change(periods=i)
             pctchange_mean = df[f"pctchange_{i}"].rolling(window=i, min_periods=i).mean()
             pctchange_std = df[f"pctchange_{i}"].rolling(window=i, min_periods=i).std()
             df[f"bb_pctchange_{i}_up"] = pctchange_mean + 2 * pctchange_std
             df[f"bb_pctchange_{i}_lo"] = pctchange_mean - 2 * pctchange_std
-            # df[f"vwrs_{i}"] = vwrs(df, length=i)
-            # df[f"close_corr_{i}"] = ef['index'].rolling(window=i, min_periods=i).corr(ef['close'], method='spearman')
-            # df[f"close_corr_{j}"] = ef['index'].rolling(window=j, min_periods=j).corr(ef['close'], method='spearman')
             df = df.copy()
         return df
 
     def populate_buy_trend(self, df: DataFrame, metadata: dict) -> DataFrame:
         df.loc[:, 'buy_tag'] = ''
         for i in self.buy_lookback_range:
-            # j = i * 2
             buy_conditions = [
                 df[f"candle_count_{self.startup_candle_count}"].ge(self.startup_candle_count),
                 df['volume'].ge(self.min_candle_vol),
@@ -201,15 +189,12 @@ class ViNBuyPct(ViN):
                 (df[f"pctchange_{i}"] / df[f"bb_pctchange_{i}_lo"]).between(1.01, 1.20),
                 (df[f"bb_pctchange_{i}_up"] - df[f"bb_pctchange_{i}_lo"]).ge(0.03),
                 (df['lc2_adj'] / df['close']).le(0.994)
-                # df[f"vwrs_{i}"].le(i + 4),
-                # df[f"close_corr_{j}"].gt(-0.80),
-                # (df[f"close_corr_{j}"] - df[f"close_corr_{i}"]).gt(0)
             ]
             buy = reduce(lambda x, y: x & y, buy_conditions)
             df.loc[buy, 'buy_tag'] += f"{i} "
         df.loc[:, 'buy_signals'] = df['buy_tag'].str.len() // 3
-        n = [1, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
-        df.loc[:, 'buy'] = df['buy_tag'].ne('') & df['buy_signals'].isin(n)
+        n = [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13]
+        df.loc[:, 'buy'] = df['buy_tag'].str.contains('25|26|27|28|29|30|31|32') | df['buy_signals'].isin(n)
         df.loc[:, 'buy_tag'] = df['buy_tag'].str.strip()
 
         self.fill_custom_buy_info(df, metadata)
@@ -224,25 +209,13 @@ class ViNSellCorrV1(ViN):
         df_trade: DataFrame = df.loc[df['date'].ge(trade_open_date - timedelta(minutes=5))]
         trade_len = len(df_trade) - 1
         candle_1 = df_trade.iloc[-1]
-        ef = df_trade[['close', 'hlc3_adj', 'volume']].reset_index()
         current_profit = (candle_1['close'] - trade.open_rate) / trade.open_rate
-        buy_tag = trade.buy_tag
-        buy_signals = (len(buy_tag) + 1) // 3
-        if trade_len <= 2:
-            if current_profit < -0.08 and not candle_1['green']:
-                log.info(f"custom_sell: immediate sell for pair {pair} with loss {current_profit:.2f} and trade len {trade_len} on candle {df_trade['date'].iat[-1]}.")
-                return f"immediate sell ({buy_signals})"
-            else:
-                return None
-        if trade_len < self.sideways_candles and -0.015 < current_profit < 0.015:
+        if trade_len <= 2 or candle_1['buy'] or candle_1['streak_b'] >= candle_1['streak_s_min'] or (trade_len < self.startup_candle_count and -0.04 < current_profit < 0.04):
             return None
-        if current_profit < -0.16 and candle_1['streak_s_max'] < 1:
-                log.info(f"custom_sell: stoploss for pair {pair} with loss {current_profit:.2f} and trade len {trade_len} on candle {df_trade['date'].iat[-1]}.")
-                return f"stoploss ({buy_signals})"
-        if candle_1['streak_b'] >= candle_1['streak_s_min']:
-            return None
+        buy_signals = (len(trade.buy_tag) + 1) // 3
         i = min(trade_len, self.startup_candle_count)
         j = i // 2
+        ef = df_trade[['close', 'hlc3_adj', 'volume']].reset_index()
         ef['vwrs_sell'] = vwrs(ef, length=i)
         ef['vwrs_corr'] = ef['index'].rolling(window=i, min_periods=1).corr(ef['vwrs_sell'], method='spearman')
         ef['close_corr_i'] = ef['index'].rolling(window=i, min_periods=1).corr(ef['close'], method='spearman')
@@ -250,19 +223,29 @@ class ViNSellCorrV1(ViN):
         vwrs_corr_diff = ef['vwrs_corr'].iat[-1] - ef['vwrs_corr'].iat[-2]
         close_corr_i_diff = ef['close_corr_i'].iat[-1] - ef['close_corr_i'].iat[-2]
         close_corr_ij_diff = ef['close_corr_i'].iat[-1] - ef['close_corr_j'].iat[-1]
-        if current_profit < -0.04:
-            offset = pow(current_profit * 100, 3) / 3600 - i / 720
+        if current_profit < -0.08:
+            offset = 0.4 + pow(current_profit * 100, 3) / 1000
+            streak_s_max_lt = 1
+            streak_s_min_lt = 0
+            t = 'big loss'
+        elif current_profit < 0:
+            offset = 0.6 - pow(current_profit * 100, 2) / 1000 - trade_len * 0.018
+            streak_s_max_lt = 1
+            streak_s_min_lt = -1
+            t = 'loss'
         else:
-            offset = (sqrt(abs(current_profit * 100)) - i / (abs(current_profit * 100) + 0.01)) / 7200
-        if close_corr_i_diff > offset and close_corr_ij_diff > 0:
-            log.info(f"custom_sell: corr sell for pair {pair} with profit/loss {current_profit:.2f} and trade len {trade_len} on candle {df_trade['date'].iat[-1]}.")
-            return f"corr sell ({buy_signals})"
-        elif ef['vwrs_sell'].iat[-1] > 82 + current_profit and vwrs_corr_diff < 0:
-            log.info(f"custom_sell: vwrs sell for pair {pair} with profit/loss {current_profit:.2f} and trade len {trade_len} on candle {df_trade['date'].iat[-1]}.")
-            return f"vwrs sell ({buy_signals})"
+            offset = -0.18 + sqrt(current_profit) / 100
+            streak_s_max_lt = 36
+            streak_s_min_lt = 18
+            t = 'profit'
+        if ef['vwrs_sell'].iat[-1] > 18 and close_corr_i_diff > offset and close_corr_ij_diff > 0 and candle_1['streak_s_max'] < streak_s_max_lt and candle_1['streak_s_min'] < streak_s_min_lt:
+            log.info(f"custom_sell: corr sell for pair {pair} with {t} {current_profit:.2f} and trade len {trade_len} on candle {candle_1['date']}.")
+            return f"corr sell {t} ({buy_signals})"
+        elif ef['vwrs_sell'].iat[-1] > 96 + abs(current_profit) * 18 - max(54, trade_len * 0.18) and vwrs_corr_diff < 0 and candle_1['hc2_adj'] / candle_1['close'] >= 1.006:
+            log.info(f"custom_sell: vwrs sell for pair {pair} with {t} {current_profit:.2f} and trade len {trade_len} on candle {candle_1['date']}.")
+            return f"vwrs sell {t} ({buy_signals})"
         else:
             return None
-
 
 class ViNBuyPctSellCorrV1(ViNBuyPct, ViNSellCorrV1):
     pass
